Fix broken test when the morpheme file does not exist a priori

+++ tests/test_create_dumps.py
@@ -6,7 +6,8 @@ from acres import create_dumps
 def test_create_morpho_dump():
     output_file = "models/pickle/morphemes.p"
 
-    os.remove(output_file)
+    if os.path.isfile(output_file):
+        os.remove(output_file)
     assert not os.path.isfile(output_file)
 
     create_dumps.create_morpho_dump()
